Backed out 4 changesets (bug 1677426) for SprintfLiteral related bustage CLOSED TREE
Backed out changeset 4c307b4c5cec (bug 1677426)
Backed out changeset 5c8089c089f1 (bug 1677426)
Backed out changeset 37ac8dd26e51 (bug 1677426)
Backed out changeset 6424b1aba83d (bug 1677426)
