+++ js/src/vm/Interpreter.h
@@ -20,7 +20,6 @@
 
 namespace js {
 
-class WithScope;
 class EnvironmentIter;
 class PlainObject;
