+++ js/src/vm/JSScript.cpp
@@ -690,92 +690,6 @@ void js::BaseScript::swapData(UniquePtr<PrivateScriptData>& other) {
   other.reset(tmp);
 }
 
-js::Scope* js::BaseScript::enclosingScope() const {
-  MOZ_ASSERT(!warmUpData_.isEnclosingScript(),
-             "Enclosing scope is not computed yet");
-
-  if (warmUpData_.isEnclosingScope()) {
-    return warmUpData_.toEnclosingScope();
-  }
-
-  MOZ_ASSERT(data_, "Script doesn't seem to be compiled");
-
-  return gcthings()[js::GCThingIndex::outermostScopeIndex()]
-      .as<Scope>()
-      .enclosing();
-}
-
-size_t JSScript::numAlwaysLiveFixedSlots() const {
-  if (bodyScope()->is<js::FunctionScope>()) {
-    return bodyScope()->as<js::FunctionScope>().nextFrameSlot();
-  }
-  if (bodyScope()->is<js::ModuleScope>()) {
-    return bodyScope()->as<js::ModuleScope>().nextFrameSlot();
-  }
-  return 0;
-}
-
-unsigned JSScript::numArgs() const {
-  if (bodyScope()->is<js::FunctionScope>()) {
-    return bodyScope()->as<js::FunctionScope>().numPositionalFormalParameters();
-  }
-  return 0;
-}
-
-bool JSScript::functionHasParameterExprs() const {
-  // Only functions have parameters.
-  js::Scope* scope = bodyScope();
-  if (!scope->is<js::FunctionScope>()) {
-    return false;
-  }
-  return scope->as<js::FunctionScope>().hasParameterExprs();
-}
-
-js::ModuleObject* JSScript::module() const {
-  if (bodyScope()->is<js::ModuleScope>()) {
-    return bodyScope()->as<js::ModuleScope>().module();
-  }
-  return nullptr;
-}
-
-bool JSScript::isGlobalCode() const {
-  return bodyScope()->is<js::GlobalScope>();
-}
-
-js::VarScope* JSScript::functionExtraBodyVarScope() const {
-  MOZ_ASSERT(functionHasExtraBodyVarScope());
-  for (JS::GCCellPtr gcThing : gcthings()) {
-    if (!gcThing.is<js::Scope>()) {
-      continue;
-    }
-    js::Scope* scope = &gcThing.as<js::Scope>();
-    if (scope->kind() == js::ScopeKind::FunctionBodyVar) {
-      return &scope->as<js::VarScope>();
-    }
-  }
-  MOZ_CRASH("Function extra body var scope not found");
-}
-
-bool JSScript::needsBodyEnvironment() const {
-  for (JS::GCCellPtr gcThing : gcthings()) {
-    if (!gcThing.is<js::Scope>()) {
-      continue;
-    }
-    js::Scope* scope = &gcThing.as<js::Scope>();
-    if (ScopeKindIsInBody(scope->kind()) && scope->hasEnvironment()) {
-      return true;
-    }
-  }
-  return false;
-}
-
-bool JSScript::isDirectEvalInFunction() const {
-  if (!isForEval()) {
-    return false;
-  }
-  return bodyScope()->hasOnChain(js::ScopeKind::Function);
-}
-
 template <XDRMode mode>
 /* static */
 XDRResult js::PrivateScriptData::XDR(XDRState<mode>* xdr, HandleScript script,
