+++ js/src/wasm/WasmJS.cpp
@@ -1815,28 +1815,11 @@ bool WasmInstanceObject::isNewborn() const {
   return getReservedSlot(INSTANCE_SLOT).isUndefined();
 }
 
-// WeakScopeMap maps from function index to js::Scope. This maps is weak
-// to avoid holding scope objects alive. The scopes are normally created
-// during debugging.
-//
-// This is defined here in order to avoid recursive dependency between
-// WasmJS.h and Scope.h.
-using WasmFunctionScopeMap =
-    JS::WeakCache<GCHashMap<uint32_t, WeakHeapPtr<WasmFunctionScope*>,
-                            DefaultHasher<uint32_t>, ZoneAllocPolicy>>;
-class WasmInstanceObject::UnspecifiedScopeMap {
- public:
-  WasmFunctionScopeMap& asWasmFunctionScopeMap() {
-    return *(WasmFunctionScopeMap*)this;
-  }
-};
-
 /* static */
 void WasmInstanceObject::finalize(JSFreeOp* fop, JSObject* obj) {
   WasmInstanceObject& instance = obj->as<WasmInstanceObject>();
   fop->delete_(obj, &instance.exports(), MemoryUse::WasmInstanceExports);
-  fop->delete_(obj, &instance.scopes().asWasmFunctionScopeMap(),
-               MemoryUse::WasmInstanceScopes);
+  fop->delete_(obj, &instance.scopes(), MemoryUse::WasmInstanceScopes);
   fop->delete_(obj, &instance.indirectGlobals(),
                MemoryUse::WasmInstanceGlobals);
   if (!instance.isNewborn()) {
@@ -1873,8 +1856,7 @@ WasmInstanceObject* WasmInstanceObject::create(
     return nullptr;
   }
 
-  UniquePtr<WasmFunctionScopeMap> scopes =
-      js::MakeUnique<WasmFunctionScopeMap>(cx->zone(), cx->zone());
+  UniquePtr<ScopeMap> scopes = js::MakeUnique<ScopeMap>(cx->zone(), cx->zone());
   if (!scopes) {
     ReportOutOfMemory(cx);
     return nullptr;
@@ -2037,8 +2019,8 @@ WasmInstanceObject::ExportMap& WasmInstanceObject::exports() const {
   return *(ExportMap*)getReservedSlot(EXPORTS_SLOT).toPrivate();
 }
 
-WasmInstanceObject::UnspecifiedScopeMap& WasmInstanceObject::scopes() const {
-  return *(UnspecifiedScopeMap*)(getReservedSlot(SCOPES_SLOT).toPrivate());
+WasmInstanceObject::ScopeMap& WasmInstanceObject::scopes() const {
+  return *(ScopeMap*)getReservedSlot(SCOPES_SLOT).toPrivate();
 }
 
 WasmInstanceObject::GlobalObjectVector& WasmInstanceObject::indirectGlobals()
@@ -2160,8 +2142,7 @@ WasmInstanceScope* WasmInstanceObject::getScope(
 /* static */
 WasmFunctionScope* WasmInstanceObject::getFunctionScope(
     JSContext* cx, HandleWasmInstanceObject instanceObj, uint32_t funcIndex) {
-  if (auto p =
-          instanceObj->scopes().asWasmFunctionScopeMap().lookup(funcIndex)) {
+  if (ScopeMap::Ptr p = instanceObj->scopes().lookup(funcIndex)) {
     return p->value();
   }
 
@@ -2177,8 +2158,7 @@ WasmFunctionScope* WasmInstanceObject::getFunctionScope(
     return nullptr;
   }
 
-  if (!instanceObj->scopes().asWasmFunctionScopeMap().putNew(funcIndex,
-                                                             funcScope)) {
+  if (!instanceObj->scopes().putNew(funcIndex, funcScope)) {
     ReportOutOfMemory(cx);
     return nullptr;
   }
